Form Validation for Application

+++ src/main/java/FairGrounds/Domain/ExpertiseProfile.java
@@ -1,6 +1,10 @@
 package FairGrounds.Domain;
 
 import javax.persistence.*;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 @Entity
 @Table (name = "EXPERTISEPROFILE")
@@ -21,7 +25,7 @@ public class ExpertiseProfile{
      * @param years - the years of expertise
      * @param application - the application associated with the current expertise
      */
-    public ExpertiseProfile(Expertise expertise, double years, Application application){
+    public ExpertiseProfile(Expertise expertise, Double years, Application application){
         this.expertise = expertise;
         this.years=years;
         this.application = application;
@@ -34,8 +38,9 @@ public class ExpertiseProfile{
      */
     public ExpertiseProfile(Expertise expertise){
         this.expertise = expertise;
-        this.years=0;
+        this.years= (double) 0;
     }
+
 
     @Id
     @Column(name="id")
@@ -51,7 +56,7 @@ public class ExpertiseProfile{
     private Application application;
 
     @Column (name = "years")
-    private double years;
+    private Double years;
 
     public Expertise getExpertise() {
         return expertise;
@@ -77,11 +82,11 @@ public class ExpertiseProfile{
         this.application = application;
     }
 
-    public double getYears() {
+    public Double getYears() {
         return years;
     }
 
-    public void setYears(double years) {
+    public void setYears(Double years) {
         this.years = years;
     }
 }
